college_data/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=2.6.27",
    "mlflow>=3.1.4",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.0",
]
